refactor: Isolate MatchOptimizer in separate struct
